search/requirements.txt
flask
pika
opentelemetry.instrumentation.flask
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp-proto-http
opentelemetry-instrumentation-requests
opentelemetry-instrumentation-flask